requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
dj-database-url==0.5.0
Django==3.0.4
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
kiwisolver==1.2.0
matplotlib==3.2.2
numpy==1.19.0
opencv-python==4.3.0.36
Pillow==7.2.0
psycopg2==2.8.5
pyparsing==2.4.7
pytesseract==0.3.3
python-dateutil==2.8.1
pytz==2020.1
regex==2020.2.20
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
